Revert some of my larger diffs to vms/vms.c.
Revert some of my larger diffs to the VMS-specific code in order to
more closely match the upstream Perl sources.

I'm going to focus on VSI C and 32-bit pointers, and set aside trying to
build using Clang until after va_copy() is working in /Pointer=32 and a few
other Clang-specific issues are resolved. It's not worth trying to build Perl
in 64-bit pointer mode when there's no practical benefit and so many system
services are still 32-bit only, such as sys$filescan().

This version of the code removes a "\n" that was being added to fwrite() calls
in pipe_mbxtofd_ast(), which doesn't seem to be necessary, or to make any
difference at all, based on test suite results. I need to review this code.

The third argument to sys$forcex() in pipe_exit_routine(), the completion code,
is passed by value, not reference. I've fixed the call, but this change also
doesn't seem to change any behavior that I can observe. It does fix a compiler
error when __NEW_STARLET is enabled, which is how I discovered the mismatch.