Set my struct _acedef sizes to the correct length.
The previous commit changed the initializers for .ace$b_size from the
correct size for a single identifier to the full length of the struct
defined in acedef.h. This caused deletes to fail with a "no such file
or directory" error.

Set .ace$b_size to the previous value to fix the regression. The next
commit will refactor the delete and rename with ACL functions to use
shared code to add and remove a full control ACE without the extra step
of querying for an existing ACE and then possibly adding it back to
the file or directory even though it was never removed.