Minimal code changes to make Clang happy.
Update some casts and use "const" where needed to compile as C++.
The code currently doesn't build in 32-bit pointer mode due to an
issue with va_list, and it doesn't run correctly in 64-bit mode.

The 64-bit Clange build appears to be failing generating the
first MMK file where miniperl calls itself as a subprocess.
I'm going to switch back to VSI C and see if the custom pipe
code can be replaced with calls to pipe() and getname().