Remove argv0 version after checking for "&".
Don't strip the version number from argv[0] until after we test for
a trailing "&" to signal that Perl should be restarted in the background.
In that case, we want to call the same version of the .exe, but otherwise,
we don't want to hardcode that version number into files that insert the
value of argv[0], such as generated Descrip.MMS files.