Don't use strcpy() in VMS safe_popen() mode check.
strcpy() isn't defined for overlapping copies, and isn't needed in the
safe_popen() mode check, because strchr() should be used to search for
'r', 'w', and 'n' (they may not be the first letter in the string).
There's no need to remove 'F' or 'W' from the mode after seeing them.