Add fastpath for running images using vfork/execv.
For system(), popen(), exec(), and related calls, VMS currently uses
lib$spawn() to run a DCL script, VMSPIPE.COM, which sets up the
input and output streams and then runs the command, all based on DCL
symbols assigned by Perl before spawning the script.

When calling Perl or another C/C++ program (MMK, etc.), it's beneficial
to use the UNIX vfork()/execv() pair because then the C RTL will
automatically pass the open file descriptors, including any pipe to
communicate with the parent, to the child. It's also more efficient
to do it this way when possible.

Add a fastpath to my_popen() to check for cases where we know that
we are running an image: the command starts with "$", the command
is a foreign command which starts with "$", or the command is a VMS
or UNIX-style file path (contains ".", starts with "/", etc.).
If it's not clear whether this might be a DCL command, or if it's
a DCL script (starts with "@"), then we use the original code path.

This new code causes more test case failures than the original code.