Disable new exec fastpath for DCL scripts.
The new exec fastpath test was failing for commands starting with
"@" and containing filespec characters. I was originally attempting
to run DCL scripts through the fast path, but VMS fails to pass the
return value correctly back to Perl. When I removed the "@" check
from the first if branch, I forgot to explicitly exclude scripts
from the fastpath branch.