Fix several bugs in new vfork/execv fastpath.
Fix crash in safe_popen() fastpath code when piping to a second
command (getredirection processing). This is the "wbF" mode.
For this case, we need to set fd 1 to the input of the second
command, and then call waitpid() on exit so that the child
isn't killed if the parent exits first. This was never an issue
for the lib$spawn() code path because its subprocesses can outlive
the parent process, unlike those spawned by vfork()/execv().

Changed Perl_do_aspawn(), Perl_do_spawn(), and Perl_do_spawn_nowait()
to call safe_popen() instead of the now-commented-out do_spawn2().

Recognize "mcr " (case-insensitive) in the first four non-whitespace
chars of a command as a fastpath we can handle.

Change the #ifdef for the fflush() wrapper to VMS_FLUSH_TO_DISK,
since it's a different type of wrapper than the socket functions
that I #ifdef'd out as VMS_WRAP_SOCKETS earlier. Hopefully this
improves performance without any data loss.

Also #ifdef out the code that checks stdin for a mailbox/pipe and
then reopens it in binary mode if so. We shouldn't have to do this.
I'll see if any of these performance changes break any unit tests.

With this commit, the test harness still has a few hangs in io/dup.t,
op/exec.t, and others, but at least I haven't seen any ACCVIO crashes
like it was doing when piping to itself, and I haven't seen any test
case regressions. In the case of io/dup.t, the parent is redirecting
stdout and stderr to a child and then calling system(), handled by
safe_popen(), and then the child is seeing "SYS$OUTPUT:.;" for both
stdout and stderr rather than inheriting the filename (and position).
I'll see if this is an OS bug that's reproducible with a C test case.